Safari-609.1.11.1/LayoutTests/fast/text-autosizing/ios/idempotentmode/idempotent-autosizing-candidate-special-cases-expected.txt

PASS elementWasAutosized('text1') is true
PASS elementWasAutosized('text2') is false
PASS elementWasAutosized('text3') is true
PASS elementWasAutosized('text4') is true
PASS elementWasAutosized('text5') is true
PASS elementWasAutosized('text6') is false
PASS successfullyParsed is true

TEST COMPLETE
Hello world

Hello world

Hello world

Hello world

Hello world

Hello world
